test(go): fold topic-split test files into the ones matching their source
A test belongs in the _test.go paired with the file it covers, so one source
file has one place to look. Two files added on this branch named themselves
after a topic instead:

wrapping_test.go covered status.Errorf and status.PublicErrorf, which live in
error.go, and needed no access to the package's internals to do it, so it drops
the external test package and joins error_test.go.

models_live_test.go covered the same supportedModels catalog as models_test.go
and was already in the same package, so it merges outright. A live test does
not earn a file of its own; the env-var skip is what separates it.
